+++ src/LuaClosure.h
@@ -12,8 +12,8 @@ public:
   virtual void VisitGC(GCVisitor& visitor);
   virtual int PropagateGC(GCVisitor& visitor);
 
-  uint8_t isC;
-  uint8_t nupvalues;
+  int isC;
+  int nupvalues;
 
   TValue* pupvals_;
   UpVal** ppupvals_;
